requirements.txt
numpy
PyQt5
pyqtgraph
pyserial
scipy
